requirements.txt
pyTelegramBotAPI
huggingface_hub
SpeechRecognition
pydub
python-dotenv
